requirements.txt
flask
flask-cors
openai
python-dotenv
fastapi
uvicorn
pydantic
